# librbd: crypto performance optimization