I tested librbd crypto overhead in the toughest condition I could, which is memstore OSDs.
Found that the call to `bufferlist::rebuild_aligned_size_and_memory` on `BlockCrypto` is really expensive (equals about half of the openssl crypto ops).

Good news that with the rebuild removed, the profiling shows that the openssl ops are the only dominant ops from the crypto side. i.e. no other overhead is seen.

For reference, here are the numbers I got for single-threaded 4MB-sized IOs:
Raw (non-encrypted) RBD image: 2189MB/s write, 2389MB/s read
RBD image with LUKS1: 1124MB/s write, 1240MB/s read
RBD image with LUKS2: 1322MB/s write, 2068MB/s read
(there was considerable deviation in results, I picked the best numbers per each case).

On the same machine, openssl speed gave:
aes-256-xts with 512 bytes block (LUKS1): 3642MB/s
aes-256-xts with 4096 bytes block (LUKS2): 4324MB/s

BTW I initially wanted to include in this PR also the enablement of the CryptoContextPool which I forgot to enable by default when I created the luks PR.
However, I noticed in profiling that the context pool actually can harm performance since it saves only a little for the relatively few calls to `get_context`, while adding an additional function call (which is not inlined) to the very hot `init_context` and `update_context` methods.