@orozery 2021-01-28
The profiling showing the inefficiency before this PR:

![crypto-profile](https://user-images.githubusercontent.com/17771355/106147875-0eafa680-6181-11eb-8daf-a558cb7e5e5b.png)