@orozery 2021-01-28
Another BTW:
planning to compare my results to QEMU-LUKS on top of RBD.
This should be interesting. I expect significant improvement, definitely for LUKS2, but maybe also LUKS1.