@orozery 2021-01-28
QEMU numbers:

QEMU-LUKS:
```
root@bm01:~/qemu# ./build/qemu-io --object secret,id=sec0,data=password --image-opts file.driver=rbd,file.pool=rbd,key-secret=sec0,file.image=foo,file.conf=/root/ceph/build/ceph.conf
qemu-io> writev -P 0xfe 0 1000M
wrote 1048576000/1048576000 bytes at offset 0
1000 MiB, 1 ops; 0:00:01.75 (572.231 MiB/sec and 0.5722 ops/sec)
```

Our LUKS1 (through QEMO-IO, using a patched qemu-rbd driver):
```
root@bm01:~/qemu/build# ./qemu-io rbd:rbd/foo:conf=/root/ceph/build/ceph.conf:format=luks1 -f raw
format: luks1
qemu-io> writev -P 0xfe 0 1000M
wrote 1048576000/1048576000 bytes at offset 0
1000 MiB, 1 ops; 00.83 sec (1.175 GiB/sec and 1.2031 ops/sec)
```

LUKS2:
```
root@bm01:~/qemu/build# ./qemu-io rbd:rbd/foo:conf=/root/ceph/build/ceph.conf:format=luks2 -f raw
format: luks2
qemu-io> writev -P 0xfe 0 1000M
wrote 1048576000/1048576000 bytes at offset 0
1000 MiB, 1 ops; 00.79 sec (1.234 GiB/sec and 1.2633 ops/sec)
```

BTW QEMU with raw unencrypted rbd got 1.49GiB/s