@dillaman 2021-01-28
> QEMU numbers:

:open_mouth: 

Wow --- that is very unexpectedly good performance for this crypto layer. Now that I am thinking about it, it would be great to have a test that verifies that QEMU's LUKS implementation can properly read the librbd LUKS data (and vice-versa).