@orozery 2021-01-28
> > QEMU numbers:
> 
> 😮
> 
> Wow --- that is very unexpectedly good performance for this crypto layer. Now that I am thinking about it, it would be great to have a test that verifies that QEMU's LUKS implementation can properly read the librbd LUKS data (and vice-versa).

We have a test (luks-encryption.sh) which verifies that our implementation is compatible with cryptsetup.
Cryptsetup being the reference implementation for LUKS, I think it's enough to conclude that we match against QEMU LUKS as well.