@orozery 2021-02-03
tried to understand why our implementation is much faster and got the following:

First, I timed gcrypt (qemu's encryption library) performance, using:
```
	err = gcry_cipher_open(&handle, GCRY_CIPHER_AES256, GCRY_CIPHER_MODE_XTS, 0);
	err = gcry_cipher_setkey(handle, key, 32);
	for (int i = 0; i < 10000000; ++i) {
		gcry_cipher_reset(handle);
		err = gcry_cipher_setiv(handle, iv, 16);
		err = gcry_cipher_encrypt(handle, out, 512, in, 512);
	}
```
And got only 839MB/s (comparing to OpenSSL speed which gave 3642 MB/s).
I guess this should explain most of it.
BTW QEMU also supports nettle, but its performance is much much worse.

Another place where are better is the number of buffer copies along the IO path.
In QEMU-LUKS over RBD you got 2 copies: client -> qemu crypto -> rbd
Whereas in using our stack you get only a single copy: client -> rbd.

This single client->rbd copy takes about 33% of the time it takes to encrypt the same buffer using OpenSSL.