@orozery 2021-02-03
looks like openssl speed does not measure re-setting IV's per each block, so I implemented my own timing and got slightly lower numbers:
512 blocks: 3200 MB/s
4096 blocks: 4308 MB/s